Implement ``__finitef()`` and ``__finite()`` internal glibc/uclibc function inside KLEE's intrinsic runtime so that its possible run the ``float_classify.c`` test case without uclibc and libm. Add a copy of the test for doubles as well.

+++ test/Floats/float_classify_double.c
@@ -8,8 +8,8 @@
 #include <assert.h>
 
 int main() {
-  float x, y, z;
-  klee_make_symbolic(&x, sizeof(float), "x");
+  double x, y, z;
+  klee_make_symbolic(&x, sizeof(double), "x");
   switch (fpclassify(x)) {
   case FP_NAN:
     printf("nan\n");
